Fix print format warnings in kernel/syscall.cpp.

+++ kernel/syscall.cpp
@@ -60,7 +60,7 @@ void Register(size_t index, void* function)
 
 	if ( SYSCALL_MAX_NUM <= index )
 		PanicF("Attempted to register system call %p to index %zu, but "
-		       "SYSCALL_MAX_NUM = %zu", function, index, SYSCALL_MAX_NUM);
+		       "SYSCALL_MAX_NUM = %zu", function, index, (size_t) SYSCALL_MAX_NUM);
 	syscall_list[index] = function;
 }
